Fixing output for only Integer Numbers

+++ src/Edge.py
@@ -3,7 +3,7 @@ class Edge(object):
         super(Edge, self).__init__()
         self.origin = origin
         self.destination = destination
-        self.distance = float(distance)
+        self.distance = int(distance)
 
     def __str__(self):
         return 'from:{}, to:{}, distance:{}'.format(self.origin, self.destination, self.distance)
